# docs/specs/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "CancelResponse": {
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "boolean"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "cancelled"
        ],
        "title": "CancelResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunDetail": {
        "properties": {
          "manifest": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest"
          },
          "run": {
            "$ref": "#/components/schemas/TaskRunRecord"
          }
        },
        "required": [
          "run"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "TaskAccepted": {
        "properties": {
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id"
        ],
        "title": "TaskAccepted",
        "type": "object"
      },
...
      "TaskRunRecord": {
        "description": "DB 行（task_runs 表的 Pydantic 映射）。\n\n与 `TaskResult` 的区别：`TaskResult` 是**一次执行的产出**（handler 返回它），\n`TaskRunRecord` 是**这次执行的档案**（含 id / 时间戳 / 进度 / 清单路径）。\n前者活在一次调用里，后者要能被\"任务历史\"页面翻出来。\n\nV1 §2 契约二看护：`status` 为终态时 `ended_at` 必填，DB 层有 CHECK 约束\n`status = 'running' OR ended_at IS NOT NULL`。",
        "properties": {
          "config_snapshot_json": {
            "default": "{}",
            "title": "Config Snapshot Json",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "error_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Text"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "manifest_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Path"
          },
          "params_json": {
            "default": "{}",
            "title": "Params Json",
            "type": "string"
          },
          "progress": {
            "default": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "success",
              "partial",
              "failed",
              "timeout",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Json"
          },
          "task_name": {
            "title": "Task Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_name",
          "kind",
          "status",
          "started_at"
        ],
        "title": "TaskRunRecord",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Intelligence Hub V2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tasks/runs": {
      "get": {
        "operationId": "list_runs_api_tasks_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRunRecord"
                  },
                  "title": "Response List Runs Api Tasks Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_tasks_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/runs/{run_id}/cancel": {
      "post": {
        "description": "协作式取消：置位 `CancelToken`。返回 False 表示这条已经不在跑了（诚实回执，不转圈）。",
        "operationId": "cancel_run_api_tasks_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/runs/{run_id}/events": {
      "get": {
        "operationId": "run_events_api_tasks_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "types",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Types"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Events",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{name}/run": {
      "post": {
        "description": "启动任务，立即返回 task_id（202）。\n\n参数校验失败 / 平台禁用 / 未实现 → `scheduler.submit` 抛 `ValidationError` / `TaskRejected`，\n交全局异常处理器翻成 422（不在这里 catch —— 那样每个入口都要抄一遍映射）。",
        "operationId": "run_task_api_tasks__name__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{name}/schema": {
      "get": {
        "operationId": "task_schema_api_tasks__name__schema_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Task Schema Api Tasks  Name  Schema Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Schema",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}